fix(fuel): adding new camps and adding new void in repository

+++ src/main/java/com/analisis/icarus/domain/Fuel/PowerSourceRepository.java
@@ -6,9 +6,9 @@ import java.util.Optional;
 public interface PowerSourceRepository {
 
     PowerSource save(PowerSource powerSource);
-    Optional<PowerSource> findById(Long id);
+    Optional<PowerSource> findById(Integer id);
     List<PowerSource> findAll();
-    void deleteById(Long id);
-    void updateById(Long id);
+    void deleteById(Integer id);
+    void updateById(Integer id);
 
 }
